+++ msd.t2_c4.springboot_restaurant/src/main/java/org.scrum/restaurant/repo/ModificareMeniuRepository.java
@@ -7,12 +7,12 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ModificareMeniuRepository extends JpaRepository<ModificareMeniu, Long> {
+public interface ModificareMeniuRepository extends JpaRepository<ModificareMeniu, Integer> {
     // Găsește modificări ale unui anumit meniu
     List<ModificareMeniu> findByidAngajat (Integer idAngajat);
 
     // Găsește modificări efectuate într-un anumit interval de timp
     List<ModificareMeniu> findByDataModificariiBetween(String startDate, String endDate);
-    List<ModificareMeniu> findByMeniu_Id(Long idMeniu);
+    List<ModificareMeniu> findByMeniu_Id(Integer idMeniu);
 
 }
